#11 refactor: SideBarItem 컴포넌트 atom -> molecule 변경
